Feature/Major: Add support for gallery moderation
This commit is much too large, but disentangling all of this stuff just doesn't seem worth it.

The database was granted a number of new columns and a new table.  This is all to support the underlying data around making galleries moderate-able, where submissions can be approved or rejected, based on a curator's verdicts.

Numerous routes were added to the API to support access to these database properties.